Simplify interface: have strfmt() format into a statically allocated buffer.
Instead of passing a buffer to the strfmt() call, have it use some statically
allocated internal buffer. Of course, this limits the use to a single
format at a time ... but that is anyway what we're doing :)

This increases readability (less parameters to pass) and is totally
fine for this embedded use-case.

Also saves 8 bytes of code :)