Add a page that shows focal lengths and ƒ-number for selected mirror sizes. Choose with button.
A button allows to rotate between three pages: the regular page showing the radius and
two pages showing the focal lengths (=radius/2) and a choice of ƒ-numbers calculated
for common mirror sizes. We show 6 different mirror sizes on two pages.

Also here, we support metric and imperial, depending on the dial setting.
Choice of metric sizes is {150mm, 200mm, 250mm, 300mm, 400mm, 600mm}, and
the corresponding imperial sizes are {6", 8", 10", 12", 16", 20"}.