New tactile button; slot-access cables; cutouts for components below.
o The old tactile button was too small and flimsy and disintegrated
  quickly. Use a new one.
o Cables are not connected anymore with JST connectors (too big), but
  instead slots are provided to solder them in. Also, some space
  above the board is provided for the cable access to minimize space-loss
  due to discomboulated cable entrance.
o Provide cutouts for places where on the 1.3" displays typically components
  are to be found.
o Make the board slightly less high as the row of passives in some displays
  starts a bit further up.

This is work-in-progress. Cleanup of silk-screen has to be done and
cut-outs need to be prettified. Also, GND is not connected everywhere yet.