Simplify UI: button cycles through mirror sizes directly, show ƒ/N on main screen.
o No pages anymore. Use the first two lines to show ƒ/N for selected mirror
  size, and ƒ value calculated from sphere (which is just r/2)
o sag value had to go. Not really needed as this can be read from the
  display of the indicator directly (this was more important in the early
  days to verify that we read what the indicator says).
o Big font size had to be reduced by one point or so to exactly fit in
  32 pixels (=4 stripes). Thus two small text-lines at the top and one
  large radius value at the bottom don't interfere.
o The mirror size is shown on the right top, where the physical button
  is pointing to.