Add a tool reference mode.
When zero-ing, we now can choose between a flat surface and 'tool reference'.

In tool reference mode, we zero the spherometer while standing on the tool,
then we can measure the mirror relative to that tool. Advantage is, that there
is no reference flat surface plate needed: just all the parts alreaday available
can be used.

This requires that the tool already has acquired the same shape as the mirror
though.